// app/types/user.ts
export interface User {
    id: number;
    firstName: string;
    lastName: string;
    maidenName?: string;
    age: number;
    gender: string;
    email: string;
    phone: string;
    username: string;
    birthDate: string;
    image: string;
    bloodGroup: string;
    height: number;
    weight: number;
    eyeColor: string;
    hair: {
        color: string;
        type: string;
    };
    ip?: string;
    address?: any;
}

export interface UsersResponse {
    users: User[];
    total: number;
    skip: number;
    limit: number;
}